classifier_service/requirements.txt
fastapi
uvicorn
requests
psycopg2-binary
python-multipart

